refactor: Move action scheduler
